+++ engine/common/model.c
@@ -1242,7 +1242,7 @@ static void Mod_LoadDeluxemap( void )
 
 	if( world.vecdatasize != world.litdatasize )
 	{
-		MsgDev( D_ERROR, "Mod_LoadDeluxemap: %s has mismatched size (%i should be %i)\n", path, world.vecdatasize, world.litdatasize );
+		MsgDev( D_ERROR, "Mod_LoadDeluxemap: %s has mismatched size (%i should be %i)\n", path, (int)world.vecdatasize, (int)world.litdatasize );
 		world.deluxedata = NULL;
 		world.vecdatasize = 0;
 		Mem_Free( in );
@@ -1798,7 +1798,7 @@ void Mod_BuildSurfacePolygons( msurface_t *surf, mextrasurf_t *info )
 	{
 		if( i == MAX_SIDE_VERTS )
 		{
-			MsgDev( D_ERROR, "BuildSurfMesh: poly %i exceeded %i vertexes!\n", surf - loadmodel->surfaces, MAX_SIDE_VERTS );
+			MsgDev( D_ERROR, "BuildSurfMesh: poly %i exceeded %i vertexes!\n", (int)(surf - loadmodel->surfaces), (int)MAX_SIDE_VERTS );
 			break; // too big polygon ?
 		}
 
@@ -1825,7 +1825,7 @@ void Mod_BuildSurfacePolygons( msurface_t *surf, mextrasurf_t *info )
 		texname = surf->texinfo->texture->name;
 	else texname = "notexture";
 
-	MsgDev( D_ERROR, "BuildSurfMesh: surface %i (%s) failed to build surfmesh\n", surf - loadmodel->surfaces, texname );
+	MsgDev( D_ERROR, "BuildSurfMesh: surface %i (%s) failed to build surfmesh\n", (int)(surf - loadmodel->surfaces), texname );
 }
 
 /*
@@ -1863,7 +1863,7 @@ void Mod_TesselatePolygon( msurface_t *surf, model_t *mod, float tessSize )
 	{
 		if( i == MAX_SIDE_VERTS )
 		{
-			MsgDev( D_ERROR, "BuildSurfMesh: poly %i exceeded %i vertexes!\n", surf - loadmodel->surfaces, MAX_SIDE_VERTS );
+			MsgDev( D_ERROR, "BuildSurfMesh: poly %i exceeded %i vertexes!\n", (int)(surf - loadmodel->surfaces), (int)MAX_SIDE_VERTS );
 			break; // too big polygon ?
 		}
 
@@ -1885,7 +1885,7 @@ void Mod_TesselatePolygon( msurface_t *surf, model_t *mod, float tessSize )
 		texname = surf->texinfo->texture->name;
 	else texname = "notexture";
 
-	MsgDev( D_ERROR, "BuildSurfMesh: surface %i (%s) failed to build surfmesh\n", surf - loadmodel->surfaces, texname );
+	MsgDev( D_ERROR, "BuildSurfMesh: surface %i (%s) failed to build surfmesh\n", (int)(surf - loadmodel->surfaces), texname );
 }
 
 /*
